Update signer demo auth layout
